% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add_prefix_to_columns}
\alias{add_prefix_to_columns}
\title{Helper Method to add a prefix to all columns (which are not excluded)}
\usage{
add_prefix_to_columns(data, prefix, ignored_column_names)
}
\arguments{
\item{data}{data.frame with columns}

\item{prefix}{string prefix to be added}

\item{ignored_column_names}{list columns to be ignored}
}
\value{
data.frame with renamed columns
}
\description{
Helper Method to add a prefix to all columns (which are not excluded)
}
\examples{
data <- add_prefix_to_columns(data, "prefix", c("id"))
}
