% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fetch_survey_data}
\alias{fetch_survey_data}
\title{Fetch Survey Data From Formr}
\usage{
fetch_survey_data(
  formr_email,
  formr_password,
  initial_survey_name,
  daily_survey_name
)
}
\arguments{
\item{formr_email}{string the mail address of the formr account}

\item{formr_password}{string the password of the formr account}

\item{initial_survey_name}{string the internal formr name of the initial survey}

\item{daily_survey_name}{string the internal formr name of the daily survey}
}
\description{
Fetch Survey Data From Formr
}
\examples{
fetch_survey_data("test@test.de", "passwort123", "initial_survey", "daily_survey")
}
