% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_daily_survey_data}
\alias{get_daily_survey_data}
\title{Function that retreives the daily survey data of an participant}
\usage{
get_daily_survey_data(id)
}
\arguments{
\item{id}{string the id of the participant}
}
\value{
data.frame with all daily survey entries from the participant
}
\description{
Function that retreives the daily survey data of an participant
}
