% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_initial_survey_data}
\alias{get_initial_survey_data}
\title{Function that retrieves initial survey data of an participant}
\usage{
get_initial_survey_data(id)
}
\arguments{
\item{id}{string session id of the participant}
}
\value{
data.frame with initial survey data
}
\description{
Function that retrieves initial survey data of an participant
}
