% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_screen_time_data_for_date}
\alias{get_screen_time_data_for_date}
\title{Helper function to get screen time data for a specific user on a specific day}
\usage{
get_screen_time_data_for_date(session, startdate, enddate)
}
\arguments{
\item{session}{string - internal formr session id}

\item{startdate}{string provided like "YYYY-mm-DD"}

\item{enddate}{string provided like "YYYY-mm-DD"}
}
\value{
data.frame with screen time data for session
}
\description{
Helper function to get screen time data for a specific user on a specific day
}
