% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_single_data}
\alias{get_single_data}
\title{Function to retrieve a single data set}
\usage{
get_single_data(session_id, rescue_time_api_key, scope, all_data)
}
\arguments{
\item{session_id}{string session id of the participant}

\item{rescue_time_api_key}{string rescue time api key of the participant}

\item{scope}{string rescue time scope (usually 'Activity')}

\item{all_data}{data.frame the data of all participants}
}
\value{
data.frame containing only the indivuduals data enriched with detailes screen time data
}
\description{
Function to retrieve a single data set
}
\examples{
all_data <- read.csv("all_data.csv")
data <- get_single_data("1", "ABC", "Activity", all_data)
write.csv(data, "single_data.csv", na = "", row.names = FALSE)
}
