% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_all_data}
\alias{get_all_data}
\title{Function that returns esm and screen time data of all survey participants}
\usage{
get_all_data(previous_result = NULL)
}
\arguments{
\item{previous_data}{data.frame already fetched before (can be left empty)}
}
\description{
Function that returns esm and screen time data of all survey participants
}
\examples{
data <- get_all_data()
write.csv(data, "all_data.csv", na = "", row.names = FALSE)
}
